style(dist/index.html): relocate HTML `<meta>` tag
HTML `<meta>` tags should be located inside the `<head>` tags.

In the 3da8bd6 commit, a `<meta>` tag was located outside `<head>` tags
at the top of the `index.html` file; this should not be legit.

This commit also make the charset setting `UTF-8` lower case, as
suggested in the Google Style Guide for HTML; please see:
https://google.github.io/styleguide/htmlcssguide.html#Encoding